FYP/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
codegen==1.0
cx-Oracle==8.1.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-rest==0.8.6
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
newsapi-python==0.2.6
odfpy==1.4.1
openpyxl==3.0.6
patterns==0.3
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
uWSGI==2.0.19.1
npm==0.1.1
optional-django==0.1.0